requirements.txt
cachetools==4.2.2
certifi==2025.10.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==22.0
pytz==2025.2
six==1.17.0
tornado==6.1
typing_extensions==4.15.0
tzlocal==5.3.1